Stop depending on Memoized.class, and move the extension+processor to its own class
RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=188329311
